requirements.txt
Flask
google-generativeai
Flask-Cors
python-dotenv
gunicorn